+++ packages/learn-card-init/src/initializers/emptyLearnCard.ts
@@ -1,4 +1,5 @@
 import { generateLearnCard } from '@learncard/core';
+import { DynamicLoaderPlugin } from '@learncard/dynamic-loader-plugin';
 import { CryptoPlugin } from '@learncard/crypto-plugin';
 import { getDidKitPlugin } from '@learncard/didkit-plugin';
 import { expirationPlugin } from '@learncard/expiration-plugin';
@@ -13,12 +14,16 @@ import { EmptyLearnCard } from 'types/LearnCard';
  *
  * @group Init Functions
  */
-export const emptyLearnCard = async ({ didkit, debug }: EmptyLearnCard['args'] = {}): Promise<
-    EmptyLearnCard['returnValue']
-> => {
-    const cryptoLc = await (await generateLearnCard({ debug })).addPlugin(CryptoPlugin);
+export const emptyLearnCard = async ({
+    didkit,
+    allowRemoteContexts = false,
+    debug,
+}: EmptyLearnCard['args'] = {}): Promise<EmptyLearnCard['returnValue']> => {
+    const dynamicLc = await (await generateLearnCard({ debug })).addPlugin(DynamicLoaderPlugin);
 
-    const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(didkit));
+    const cryptoLc = await dynamicLc.addPlugin(CryptoPlugin);
+
+    const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(didkit, allowRemoteContexts));
 
     const expirationLc = await didkitLc.addPlugin(expirationPlugin(didkitLc));