+++ packages/learn-card-init/src/initializers/learnCardFromSeed.ts
@@ -1,4 +1,5 @@
 import { generateLearnCard } from '@learncard/core';
+import { DynamicLoaderPlugin } from '@learncard/dynamic-loader-plugin';
 import { CryptoPlugin } from '@learncard/crypto-plugin';
 import { DidMethod, getDidKitPlugin } from '@learncard/didkit-plugin';
 import { getDidKeyPlugin } from '@learncard/didkey-plugin';
@@ -31,12 +32,15 @@ export const learnCardFromSeed = async ({
     } = {},
     ceramicIdx = defaultCeramicIDXArgs,
     didkit,
+    allowRemoteContexts = false,
     ethereumConfig = defaultEthereumArgs,
     debug,
 }: LearnCardFromSeed['args']): Promise<LearnCardFromSeed['returnValue']> => {
-    const cryptoLc = await (await generateLearnCard({ debug })).addPlugin(CryptoPlugin);
+    const dynamicLc = await (await generateLearnCard({ debug })).addPlugin(DynamicLoaderPlugin);
 
-    const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(didkit));
+    const cryptoLc = await dynamicLc.addPlugin(CryptoPlugin);
+
+    const didkitLc = await cryptoLc.addPlugin(await getDidKitPlugin(didkit, allowRemoteContexts));
 
     const didkeyLc = await didkitLc.addPlugin(
         await getDidKeyPlugin<DidMethod>(didkitLc, seed, 'key')