+++ packages/plugins/learn-card/src/verify.ts
@@ -14,9 +14,9 @@ const transformErrorMessage = (error: string, credential: VC): string => {
     if (error.startsWith('expiration')) {
         return credential.expirationDate
             ? `Invalid • Expired ${format(
-                  new Date(credential.expirationDate),
-                  'dd MMM yyyy'
-              ).toUpperCase()}`
+                new Date(credential.expirationDate),
+                'dd MMM yyyy'
+            ).toUpperCase()}`
             : 'Invalid • Expired';
     }
 
@@ -29,9 +29,9 @@ const transformCheckMessage = (check: string, credential: VC): string => {
             proof: 'Valid',
             expiration: credential.expirationDate
                 ? `Valid • Expires ${format(
-                      new Date(credential.expirationDate),
-                      'dd MMM yyyy'
-                  ).toUpperCase()}`
+                    new Date(credential.expirationDate),
+                    'dd MMM yyyy'
+                ).toUpperCase()}`
                 : 'Valid • Does Not Expire',
         }[check] || check
     );