+++ packages/react-learn-card/src/components/VCDisplayCard2/VC2FrontFaceInfo.tsx
@@ -10,6 +10,7 @@ type VC2FrontFaceInfoProps = {
     issuee: Profile | string;
     issuer: Profile | string;
     title: string;
+    subjectDID?: string | undefined;
     subjectImageComponent?: React.ReactNode;
     issuerImageComponent?: React.ReactNode;
     createdAt: string;
@@ -21,6 +22,7 @@ type VC2FrontFaceInfoProps = {
 const VC2FrontFaceInfo: React.FC<VC2FrontFaceInfoProps> = ({
     issuee,
     issuer,
+    subjectDID,
     subjectImageComponent,
     issuerImageComponent,
     customBodyCardComponent,
@@ -72,8 +74,13 @@ const VC2FrontFaceInfo: React.FC<VC2FrontFaceInfoProps> = ({
 
                 {!customBodyCardComponent && (
                     <>
-                        <h3 className="text-[27px] leading-[130%] text-grayscale-900">
+                        <h3 className="text-[27px] flex flex-col text-center leading-[130%] text-grayscale-900 capitalize">
                             {issueeName}
+                            {subjectDID && (
+                                <span className="text-[12px] text-grayscale-700 leading-[18px] font-poppins font-[400] m-0 p-0 normal-case">
+                                    {subjectDID}
+                                </span>
+                            )}
                         </h3>
                         <div className="relative">
                             <div className="vc-issuee-image h-[60px] w-[60px] rounded-full overflow-hidden">