+++ packages/learn-card-core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # learn-card-core
 
+## 9.1.1
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/helpers@1.0.8
+
+## 9.1.0
+
+### Minor Changes
+
+-   [#368](https://github.com/learningeconomy/LearnCard/pull/368) [`1c4e09d`](https://github.com/learningeconomy/LearnCard/commit/1c4e09d136464286959000e5ed14cdf59dba9196) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Create Context Plane
+
 ## 9.0.3
 
 ### Patch Changes