+++ packages/learn-card-init/CHANGELOG.md
@@ -1,5 +1,109 @@
 # learn-card-core
 
+## 1.2.2
+
+### Patch Changes
+
+-   Updated dependencies [[`e0c2e3c`](https://github.com/learningeconomy/LearnCard/commit/e0c2e3cb202697b6875c729e35326d75bbae4f65)]:
+    -   @learncard/didkit-plugin@1.3.2
+    -   @learncard/ceramic-plugin@1.0.5
+    -   @learncard/chapi-plugin@1.0.15
+    -   @learncard/did-web-plugin@1.0.20
+    -   @learncard/learn-card-plugin@1.1.2
+    -   @learncard/network-plugin@1.5.2
+    -   @learncard/learn-cloud-plugin@1.1.18
+    -   @learncard/vc-plugin@1.1.2
+    -   @learncard/expiration-plugin@1.1.2
+
+## 1.2.1
+
+### Patch Changes
+
+-   Updated dependencies [[`56aef2d`](https://github.com/learningeconomy/LearnCard/commit/56aef2d7830a5c66fa3b569b3c25eb3ecb6cc465)]:
+    -   @learncard/types@5.3.3
+    -   @learncard/core@9.1.1
+    -   @learncard/helpers@1.0.8
+    -   @learncard/ceramic-plugin@1.0.5
+    -   @learncard/chapi-plugin@1.0.14
+    -   @learncard/did-web-plugin@1.0.19
+    -   @learncard/didkey-plugin@1.0.5
+    -   @learncard/didkit-plugin@1.3.1
+    -   @learncard/idx-plugin@1.0.5
+    -   @learncard/learn-card-plugin@1.1.1
+    -   @learncard/network-plugin@1.5.1
+    -   @learncard/learn-cloud-plugin@1.1.17
+    -   @learncard/vc-plugin@1.1.1
+    -   @learncard/vc-api-plugin@1.0.5
+    -   @learncard/vc-templates-plugin@1.0.8
+    -   @learncard/vpqr-plugin@1.0.5
+    -   @learncard/expiration-plugin@1.1.1
+    -   @learncard/crypto-plugin@1.0.5
+    -   @learncard/dynamic-loader-plugin@1.0.1
+    -   @learncard/ethereum-plugin@1.0.5
+
+## 1.2.0
+
+### Minor Changes
+
+-   [#340](https://github.com/learningeconomy/LearnCard/pull/340) [`867d38c`](https://github.com/learningeconomy/LearnCard/commit/867d38c5f606ff73fa328a4616a8a58a8f49d0f0) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Add support for credentialSchema checks
+
+### Patch Changes
+
+-   Updated dependencies [[`867d38c`](https://github.com/learningeconomy/LearnCard/commit/867d38c5f606ff73fa328a4616a8a58a8f49d0f0)]:
+    -   @learncard/network-plugin@1.5.0
+    -   @learncard/expiration-plugin@1.1.0
+    -   @learncard/learn-card-plugin@1.1.0
+    -   @learncard/didkit-plugin@1.3.0
+    -   @learncard/vc-plugin@1.1.0
+    -   @learncard/learn-cloud-plugin@1.1.16
+    -   @learncard/did-web-plugin@1.0.18
+    -   @learncard/ceramic-plugin@1.0.4
+    -   @learncard/chapi-plugin@1.0.13
+
+## 1.1.0
+
+### Minor Changes
+
+-   [#368](https://github.com/learningeconomy/LearnCard/pull/368) [`1c4e09d`](https://github.com/learningeconomy/LearnCard/commit/1c4e09d136464286959000e5ed14cdf59dba9196) Thanks [@TaylorBeeston](https://github.com/TaylorBeeston)! - Add option for remote context resolution and bring in Dynamic Loader plugin
+
+### Patch Changes
+
+-   Updated dependencies [[`1c4e09d`](https://github.com/learningeconomy/LearnCard/commit/1c4e09d136464286959000e5ed14cdf59dba9196), [`1c4e09d`](https://github.com/learningeconomy/LearnCard/commit/1c4e09d136464286959000e5ed14cdf59dba9196), [`1c4e09d`](https://github.com/learningeconomy/LearnCard/commit/1c4e09d136464286959000e5ed14cdf59dba9196)]:
+    -   @learncard/core@9.1.0
+    -   @learncard/didkit-plugin@1.2.0
+    -   @learncard/dynamic-loader-plugin@1.0.0
+    -   @learncard/ceramic-plugin@1.0.4
+    -   @learncard/chapi-plugin@1.0.12
+    -   @learncard/crypto-plugin@1.0.4
+    -   @learncard/did-web-plugin@1.0.17
+    -   @learncard/didkey-plugin@1.0.4
+    -   @learncard/ethereum-plugin@1.0.4
+    -   @learncard/expiration-plugin@1.0.13
+    -   @learncard/idx-plugin@1.0.4
+    -   @learncard/learn-card-plugin@1.0.13
+    -   @learncard/network-plugin@1.4.33
+    -   @learncard/learn-cloud-plugin@1.1.15
+    -   @learncard/vc-plugin@1.0.13
+    -   @learncard/vc-api-plugin@1.0.4
+    -   @learncard/vc-templates-plugin@1.0.7
+    -   @learncard/vpqr-plugin@1.0.4
+
+## 1.0.23
+
+### Patch Changes
+
+-   Updated dependencies [[`5e102ff`](https://github.com/learningeconomy/LearnCard/commit/5e102ff8e41d171e26d1a070e2ca6b25a202845e)]:
+    -   @learncard/didkit-plugin@1.1.2
+    -   @learncard/vc-templates-plugin@1.0.6
+    -   @learncard/vc-plugin@1.0.12
+    -   @learncard/ceramic-plugin@1.0.3
+    -   @learncard/chapi-plugin@1.0.11
+    -   @learncard/did-web-plugin@1.0.16
+    -   @learncard/learn-card-plugin@1.0.12
+    -   @learncard/network-plugin@1.4.32
+    -   @learncard/learn-cloud-plugin@1.1.14
+    -   @learncard/expiration-plugin@1.0.12
+
 ## 1.0.22
 
 ### Patch Changes