+++ packages/learn-card-network/brain-client/CHANGELOG.md
@@ -1,5 +1,40 @@
 # @learncard/network-brain-client
 
+## 1.2.11
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-brain-service@1.6.2
+
+## 1.2.10
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-brain-service@1.6.1
+
+## 1.2.9
+
+### Patch Changes
+
+-   Updated dependencies [[`867d38c`](https://github.com/learningeconomy/LearnCard/commit/867d38c5f606ff73fa328a4616a8a58a8f49d0f0)]:
+    -   @learncard/network-brain-service@1.6.0
+
+## 1.2.8
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-brain-service@1.5.8
+
+## 1.2.7
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/network-brain-service@1.5.7
+
 ## 1.2.6
 
 ### Patch Changes