+++ packages/react-learn-card/CHANGELOG.md
@@ -1,5 +1,52 @@
 # @welibraryos/react-learn-card
 
+## 2.6.34
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.2.2
+
+## 2.6.33
+
+### Patch Changes
+
+-   [#376](https://github.com/learningeconomy/LearnCard/pull/376) [`b419f1f`](https://github.com/learningeconomy/LearnCard/commit/b419f1f4eaf229f54555ee9c60495a2edfcad07e) Thanks [@gerardopar](https://github.com/gerardopar)! - add prop to display subject did
+
+## 2.6.32
+
+### Patch Changes
+
+-   [#374](https://github.com/learningeconomy/LearnCard/pull/374) [`db5fcaf`](https://github.com/learningeconomy/LearnCard/commit/db5fcaf2afaf24d3cfb3d47219ad777cdf96108d) Thanks [@gerardopar](https://github.com/gerardopar)! - VC Backface Overrides
+
+## 2.6.31
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.2.1
+
+## 2.6.30
+
+### Patch Changes
+
+-   Updated dependencies [[`867d38c`](https://github.com/learningeconomy/LearnCard/commit/867d38c5f606ff73fa328a4616a8a58a8f49d0f0)]:
+    -   @learncard/init@1.2.0
+
+## 2.6.29
+
+### Patch Changes
+
+-   Updated dependencies [[`1c4e09d`](https://github.com/learningeconomy/LearnCard/commit/1c4e09d136464286959000e5ed14cdf59dba9196)]:
+    -   @learncard/init@1.1.0
+
+## 2.6.28
+
+### Patch Changes
+
+-   Updated dependencies []:
+    -   @learncard/init@1.0.23
+
 ## 2.6.27
 
 ### Patch Changes